+++ src/randomWalk/WalkingController.java
@@ -5,7 +5,7 @@ import java.util.Random;
 public class WalkingController {
 	private final long numberOfTrials = 10000;
 
-	public double getAverageScore(double x, double y){
+	public double getAverageScore(int x, int y){
 		double score = 0;
 
 		for(int round = 0; round<numberOfTrials; round++){
@@ -15,7 +15,7 @@ public class WalkingController {
 		return score/numberOfTrials;
 	}
 
-	private double walk(double x, double y){
+	private double walk(int x, int y){
 		Random random = new Random();
 		D2MovingPoint point = new D2MovingPoint(x, y);
 		int direction;
